EmailAgent/requirements.txt
# Core
requests==2.32.5
python-dotenv==1.2.1

# LangGraph
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langchain-core==1.2.7


# Google Auth / Gmail API
google-auth==2.47.0
google-auth-oauthlib==1.2.4
google-auth-httplib2==0.3.0
google-api-python-client==2.188.0

# Gemini / Google GenAI
google-genai==1.59.0


# Superbase
supabase==2.27.2